+++ CHANGELOG.md
@@ -15,9 +15,15 @@ files were added for all supported boards and all of their source files were reg
 usable with both *Python 2* and *Python 3*.
 - Test of `Thread::generateSignal()` and `Thread::queueSignal()` returning `ENOSPC` when the amount of target thread's
 free stack is too small to request signal delivery.
-- *GDB* pretty-printers for `estd` containers: `estd::ContiguousRange`, `estd::IntrusiveForwardList`,
-`estd::IntrusiveList`, `estd::SortedIntrusiveForwardList` and `estd::SortedIntrusiveList`. This greatly simplifies
-system debugging by allowing easy debugger access to thread lists inside scheduler and synchronization objects.
+- *GDB* pretty-printers for `estd` containers (`estd::ContiguousRange`, `estd::IntrusiveForwardList`,
+`estd::IntrusiveList`, `estd::SortedIntrusiveForwardList` and `estd::SortedIntrusiveList`),
+`distortos::internal::ThreadList` and all variants of FIFO and message queues (`distortos::internal::FifoQueueBase`,
+`distortos::FifoQueue`, `distortos::DynamicFifoQueue`, `distortos::StaticFifoQueue`, `distortos::RawFifoQueue`,
+`distortos::DynamicRawFifoQueue`, `distortos::StaticRawFifoQueue`, `distortos::internal::MessageQueueBase`,
+`distortos::MessageQueue`, `distortos::DynamicMessageQueue`, `distortos::StaticMessageQueue`,
+`distortos::RawMessageQueue`, `distortos::DynamicRawMessageQueue` and `distortos::StaticRawMessageQueue`). This greatly
+simplifies system debugging by allowing easy debugger access to thread lists (inside scheduler or synchronization
+objects), software timer lists (inside scheduler) and by allowing live examination of queue contents.
 
 ### Changed